refactor: Issue #66: applescript.rs モジュール分割
ドメイン別分割戦略を用いて1310行のmonolithicなファイルを
6つの機能別モジュールに分割しました：

- osascript.rs（33行）：osascript実行基盤
- app.rs（363行）：アプリケーション操作
- window.rs（565行）：ウィンドウ操作
- display.rs（231行）：ディスプレイ情報取得
- utils.rs（374行）：パース・ユーティリティ関数
- mod.rs（15行）：公開APIのre-export

改善点：
- 責務の明確化：各ファイルが単一の責務を持つ
- コード管理性向上：各モジュール200-500行で管理しやすい
- 拡張性向上：Windows対応など将来のマルチプラットフォーム対応に有利
- テストへの影響なし：既存テスト変更不要（APIは mod.rs で re-export）

テスト結果：
- 合計412テストパス、114テストスキップ
- cargo build：成功
- cargo test：全テストパス
